排版

+++ Making/CoreTextModel.h
@@ -22,7 +22,7 @@
 @property (nonatomic, assign) BOOL underLineForLink;       //链接是否带下划线
 @property (nonatomic, assign) BOOL autoDetectLinks;        //自动检测
 @property (nonatomic, assign) NSInteger numberOfLines;      //行数
-@property (nonatomic, assign) CGFloat mainPosition;
+@property (nonatomic, assign) NSInteger mainPosition;
 @property (nonatomic, assign) CTTextAlignment textAlignment;  //文字排版样式
 @property (nonatomic, assign) CTLineBreakMode lineBreakMode;  //LineBreakMode
 @property (nonatomic, assign) CGFloat lineSpacing;            //行间距
@@ -33,7 +33,7 @@
 @property (nonatomic, retain) NSString *forewordText;
 @property (nonatomic, assign) CGFloat forewordFontSize;
 @property (nonatomic, assign) CTTextAlignment forewordAlignment;
-@property (nonatomic, assign) CGFloat forewordPosition;
+@property (nonatomic, assign) NSInteger forewordPosition;
 @property (nonatomic, retain) NSString *forewordFontName;
 
 - (id)initWithModel:(CoreTextModel *)model;
